将业务逻辑判断移动至service层，新增所有数据字典javascript对象查询

+++ admin/admin-service/src/main/java/com/deyatech/admin/controller/DictionaryIndexController.java
@@ -1,10 +1,7 @@
 package com.deyatech.admin.controller;
 
-import cn.hutool.core.util.StrUtil;
 import cn.hutool.json.JSONUtil;
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
-import com.deyatech.admin.entity.Dictionary;
 import com.deyatech.admin.entity.DictionaryIndex;
 import com.deyatech.admin.service.DictionaryIndexService;
 import com.deyatech.admin.service.DictionaryService;
@@ -18,7 +15,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
@@ -50,21 +46,8 @@ public class DictionaryIndexController extends BaseController {
     @ApiImplicitParam(name = "dictionaryIndex", value = "系统数据字典索引信息对象", required = true, dataType = "DictionaryIndex", paramType = "query")
     public RestResult<Boolean> saveOrUpdate(DictionaryIndex dictionaryIndex) {
         log.info(String.format("保存或者更新系统数据字典索引信息: %s ", JSONUtil.toJsonStr(dictionaryIndex)));
-        if(StrUtil.isEmpty(dictionaryIndex.getId())){
-            List<DictionaryIndex> list = dictionaryIndexService.getBaseMapper().selectList(new QueryWrapper<DictionaryIndex>()
-                    .eq("key_",dictionaryIndex.getKey())
-                    .eq("enable_",1));
-            if(!list.isEmpty()){
-                return RestResult.error("已存在相同名称的索引，请重新提交！");
-            }
-        }else{
-            List<DictionaryIndex> list = dictionaryIndexService.getBaseMapper().selectList(new QueryWrapper<DictionaryIndex>()
-                    .eq("key_",dictionaryIndex.getKey())
-                    .eq("enable_",1)
-                    .notIn("id_",dictionaryIndex.getId()));
-            if(!list.isEmpty()){
-                return RestResult.error("已存在相同名称的索引，请重新提交！");
-            }
+        if(!dictionaryIndexService.validataByKey(dictionaryIndex)){
+            return RestResult.error("已存在相同的字典索引，请重新提交！");
         }
         boolean result = dictionaryIndexService.saveOrUpdate(dictionaryIndex);
         return RestResult.ok(result);
@@ -112,19 +95,9 @@ public class DictionaryIndexController extends BaseController {
     @ApiImplicitParam(name = "ids", value = "系统数据字典索引信息对象ID集合", required = true, allowMultiple = true, dataType = "Serializable", paramType = "query")
     public RestResult<Boolean> removeByIds(@RequestParam(value="ids[]") List<String> ids) {
         log.info(String.format("根据id批量删除系统数据字典索引信息: %s ", JSONUtil.toJsonStr(ids)));
-        List<DictionaryIndex> list = (List<DictionaryIndex>) dictionaryIndexService.listByIds(ids);
-        if(!list.isEmpty()){
-            List<String> indexIds = new ArrayList<>();
-            for(DictionaryIndex dictionaryIndex:list){
-                indexIds.add(dictionaryIndex.getKey());
-            }
-            boolean b = dictionaryService.remove(new QueryWrapper<Dictionary>()
-                    .in("index_id",indexIds));
-            if(!b){
-                return RestResult.error("删除出错，请联系管理员");
-            }
+        if(!dictionaryService.removeByIndexIds(ids)){
+            return RestResult.error("删除错误，请联系管理员！");
         }
-
         boolean result = dictionaryIndexService.removeByIds(ids);
         return RestResult.ok(result);
     }
